requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
audioop-lts==0.2.2
babel==2.17.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
courlan==1.3.2
dateparser==1.2.2
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.20.1
fsspec==2025.12.0
gradio==6.1.0
gradio_client==2.0.1
greenlet==3.3.0
groovy==0.1.2
h11==0.16.0
html5lib==1.1
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
openai==2.13.0
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pillow==12.0.0
pinecone==8.0.0
pinecone-client==6.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
playwright==1.57.0
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tld==0.13.1
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
validators==0.35.0
webencodings==0.5.1
wheel==0.45.1
